fix(chat): 修复气泡窗口高度和标题栏紧凑化
- 修复气泡窗口可占据90%屏幕的问题，限制最大高度为屏幕50%
- 标题栏padding从vertical:8减为4，图标缩小为20并收紧间距
- 拖拽高度比例上限从0.7降为0.5，保持30%-50%范围

+++ apps/knode_app/lib/widgets/floating_chat_ball.dart
@@ -426,7 +426,7 @@ class _PanelOverlayWidget extends StatelessWidget {
     // 面板最大高度：从 panelTop 向上到屏幕顶部，留 16px 边距
     final maxPanelHeight = panelTop - 16.0;
     // 面板高度：屏幕高度的 50%，但不超过 maxPanelHeight
-    final panelHeight = (screenHeight * 0.5).clamp(200.0, maxPanelHeight);
+    final panelHeight = (screenHeight * 0.5).clamp(200.0, maxPanelHeight).toDouble();
 
     // 面板实际顶部：从 panelTop 向上展开
     final actualTop = panelTop - panelHeight;
